Moved interface to somewhere that does not need interface to be activated.
